Refactor the directory structure (this touches almost everything) (#52)
The current content structure is confusing and does not follow good hugo practice. This touches a lot of files (485). There are three kinds of changes (see also the more detailed comment in the PR):

1. The config file now resides in `hugo.toml` (before: `config.toml`). There were only minor changes in the config (e.g., regarding the `disqus` integration).
2. All images in the blog posts were moved to the more canonical location + all links were modified accordingly.
3. The `.md` files with the actual blog posts were moved to the folder of the respective blog post, where they are simply called `index.md`.

NOTE: `git diff` managed to identify most of the moves correctly as moves, but for some of the more recent blog posts, a large diff was shown (despite an identical or at least very similar contents, apart from the mentioned image links). Not sure whether this is due to `git diff` not being able to figure out the move correctly or whether there were other reasons. But let's merge this anyway.